refactor: allow solve_fe docs without sfepy
Allow solve_fe to display its documentation without having to import
sfepy.
