+++ pymks/__init__.py
@@ -26,20 +26,8 @@ from .fmks.correlations import FlattenTransformer
 from .fmks.correlations import TwoPointCorrelation
 from .fmks.data.checkerboard import generate_checkerboard
 from .fmks.pair_correlations import paircorr_from_twopoint
+from .fmks.data import solve_fe
 
-try:
-    import sfepy  # noqa: F401
-except ImportError:
-
-    def solve_fe(*_, **__):
-        """Dummy funcion when sfepy unavailable
-        """
-        # pylint: disable=redefined-outer-name, import-outside-toplevel, unused-import
-        import sfepy  # noqa: F401, F811
-
-
-else:
-    from .fmks.data.elastic_fe import solve_fe
 
 # the following will be deprecated
 from .mks_localization_model import MKSLocalizationModel
@@ -51,7 +39,7 @@ from .mks_homogenization_model import MKSHomogenizationModel
 MKSRegressionModel = MKSLocalizationModel
 DiscreteIndicatorBasis = PrimitiveBasis
 ContinuousIndicatorBasis = PrimitiveBasis
-# the above will be deprecatec
+# the above will be deprecated
 
 
 def test():
